refactor: card totalizer flow adapted
