+++ frontend/src/pages/cards/totalizers/CardTotalizer.styles.ts
